+++ e2e/cypress.config.js
@@ -1,0 +1,16 @@
+const { defineConfig } = require("cypress");
+
+module.exports = defineConfig({
+  e2e: {
+    setupNodeEvents(on, config) {
+      on("task", {
+        log(message) {
+          // eslint-disable-next-line no-console
+          console.log(message);
+          return null;
+        },
+      });
+      return require("./cypress/plugins/index")(on, config);
+    },
+  },
+});