+++ e2e/cypress/e2e/lms_problem_interaction/latex_problem_hint.cy.js
@@ -1,31 +1,31 @@
-// LMS Multiple Choice Problem Interaction Test
+// LMS LaTeX Problem With Hint Interaction Test
 
 import { getProblem, getSectionAndURL, getXblockId } from "../../support/utils";
 
-describe("LMS Multiple Choice Problem Interaction Test", () => {
-  const [section, sectionUrl] = getSectionAndURL("multiplechoice");
-  const problem = getProblem(section, "multiplechoice");
+describe("LMS LaTeX Problem With Hint Interaction Test", () => {
+  const [section, sectionUrl] = getSectionAndURL("problemWithHintInLatex");
+  const problem = getProblem(section, "problemWithHintInLatex");
   const problemId = getXblockId(problem);
 
   before(() => {
-    cy.lmsLoginStudent();
-    cy.lmsEnroll(true);
+    cy.lmsCreateUser().then(({ email, password }) => {
+      cy.lmsLogin(email, password);
+      cy.lmsEnroll(true);
+    });
     // Navigate to the courseware.
     cy.visit(sectionUrl);
     // Input wrong answer.
-    cy.get(`#input_${problemId}_2_1_choice_brazil`).check();
+    cy.get(`#input_${problemId}_2_1`).clear().type("java");
     // Submit answer.
     cy.get(".check.Valider").click();
     cy.get(".check.Valider").should("not.have.class", "is-disabled");
-    cy.get(".hint-label").should("contain", "Incorrect");
+    cy.get(`#status_${problemId}_2_1`).should("contain", "incorrect");
     // Input correct answer.
-    const indonesiaInput = `#input_${problemId}_2_1_choice_indonesia`;
-    cy.get(indonesiaInput).check();
+    cy.get(`#input_${problemId}_2_1`).clear().type("python");
     // Submit answer.
     cy.get(".check.Valider").click();
     cy.get(".check.Valider").should("not.have.class", "is-disabled");
-    cy.get(`${indonesiaInput} + span`).should("contain", "correct");
-    cy.lmsEnroll(false);
+    cy.get(`#status_${problemId}_2_1`).should("contain", "correct");
   });
 
   it("should log problem_check server event", () => {