+++ e2e/cypress/e2e/lms_problem_interaction/option_response.cy.js
@@ -1,38 +1,31 @@
-// LMS Checkboxes Response Problem Interaction Test
+// LMS Option Response Problem Interaction Test
 
 import { getProblem, getSectionAndURL, getXblockId } from "../../support/utils";
 
-describe("LMS Checkboxes Response Interaction Test", () => {
-  const [section, sectionUrl] = getSectionAndURL("checkboxesResponse");
-  const problem = getProblem(section, "checkboxesResponse");
+describe("LMS Option Response Problem Interaction Test", () => {
+  const [section, sectionUrl] = getSectionAndURL("optionresponse");
+  const problem = getProblem(section, "optionresponse");
   const problemId = getXblockId(problem);
 
   before(() => {
-    cy.lmsLoginStudent();
-    cy.lmsEnroll(true);
+    cy.lmsCreateUser().then(({ email, password }) => {
+      cy.lmsLogin(email, password);
+      cy.lmsEnroll(true);
+    });
     // Navigate to the courseware.
     cy.visit(sectionUrl);
-    // Input wrong answers.
-    cy.get(`#input_${problemId}_2_1_choice_0`).check();
-    cy.get(`#input_${problemId}_2_1_choice_1`).check();
-    cy.get(`#input_${problemId}_2_1_choice_2`).check();
-    cy.get(`#input_${problemId}_2_1_choice_3`).check();
-    cy.get(`#input_${problemId}_2_1_choice_4`).check();
+    // Input wrong answer.
+    cy.get(`#input_${problemId}_2_1`).select("China");
     // Submit answer.
     cy.get(".check.Valider").click();
     cy.get(".check.Valider").should("not.have.class", "is-disabled");
     cy.get(`#status_${problemId}_2_1`).should("contain", "incorrect");
     // Input correct answers.
-    cy.get(`#input_${problemId}_2_1_choice_0`).check();
-    cy.get(`#input_${problemId}_2_1_choice_1`).uncheck();
-    cy.get(`#input_${problemId}_2_1_choice_2`).check();
-    cy.get(`#input_${problemId}_2_1_choice_3`).check();
-    cy.get(`#input_${problemId}_2_1_choice_4`).uncheck();
+    cy.get(`#input_${problemId}_2_1`).select("India");
     // Submit answer.
     cy.get(".check.Valider").click();
     cy.get(".check.Valider").should("not.have.class", "is-disabled");
     cy.get(`#status_${problemId}_2_1`).should("contain", "correct");
-    cy.lmsEnroll(false);
   });
 
   it("should log problem_check server event", () => {